AG-9758 - Address PR review feedback
- Add ascending-order validation to colorScale.domain
- Move pure color scale utility functions to ag-charts-core for tree-shaking
- Add enclosing describe block to stops.test.ts
- Add v13.3.0 to @deprecated colorRange tags
- Add agentic rule for module-support.ts barrel export guidelines
