+++ packages/ag-charts-types/src/series/standalone/sunburstOptions.ts
@@ -65,7 +65,7 @@ export interface AgSunburstSeriesThemeableOptions<TDatum = DatumDefault, TContex
     strokeWidth?: PixelSize;
     /**
      * The colour range to interpolate the numeric colour domain (min and max `colorKey` values) into.
-     * @deprecated Use `colorScale.fills` instead.
+     * @deprecated v13.3.0 Use `colorScale.fills` instead.
      */
     colorRange?: CssColor[];
     /** Configuration for colour scale with fills, domain, and mode. */
