+++ packages/ag-charts-types/src/series/topology/mapMarkerOptions.ts
@@ -92,7 +92,7 @@ export interface AgMapMarkerSeriesThemeableOptions<TDatum = DatumDefault, TConte
     sizeDomain?: [number, number];
     /**
      * The colour range to interpolate the numeric colour domain (min and max `colorKey` values) into.
-     * @deprecated Use `colorScale.fills` instead.
+     * @deprecated v13.3.0 Use `colorScale.fills` instead.
      */
     colorRange?: CssColor[];
     /** Configuration for colour scale with fills, domain, and mode. */
